AutomatedMonitoring/NetworkMonitor/requirements.txt
# Core dependencies
networkx>=2.5.1         # For topology mapping and graph operations
requests>=2.26.0        # For HTTP/API operations
jinja2>=3.0.1          # For template rendering
pysnmp>=4.4.12         # For SNMP operations
scapy>=2.4.5           # For network packet operations
dnspython>=2.1.0       # For DNS health checks
python-ldap>=3.3.1     # For Active Directory operations

# Network monitoring
netmiko>=3.4.0         # For network device CLI interactions
napalm>=3.3.1          # For multi-vendor network automation
nornir>=3.1.1          # For network automation framework
paramiko>=2.8.0        # SSH client for device connections
textfsm>=1.1.2         # For parsing semi-structured text
ntc-templates>=2.0.0   # Network to Code TextFSM templates

# Data processing
pandas>=1.3.0          # For data analysis and reporting
numpy>=1.21.0          # For numerical operations
influxdb-client>=1.21.0  # For time-series data storage
prometheus-client>=0.11.0  # For metrics export

# Visualization
plotly>=5.1.0          # For interactive graphs
dash>=2.0.0            # For web dashboards
graphviz>=0.17         # For network topology visualization

# Web and API
fastapi>=0.68.0        # For REST API
uvicorn>=0.15.0        # ASGI server for FastAPI
aiohttp>=3.7.4         # Async HTTP client/server
websockets>=9.1        # For real-time updates

# Security
cryptography>=3.4.7    # For encryption operations
pyOpenSSL>=20.0.1      # For SSL/TLS operations
bcrypt>=3.2.0          # For password hashing

# Email and notifications
aiosmtplib>=1.1.6      # Async SMTP client
slack-sdk>=3.11.2      # For Slack integration
python-telegram-bot>=13.7  # For Telegram integration

# Testing and development
pytest>=6.2.5
pytest-asyncio>=0.15.1
pytest-cov>=2.12.1
black>=21.7b0          # Code formatting
isort>=5.9.3          # Import sorting
flake8>=3.9.2         # Code linting
mypy>=0.910           # Static type checking

# Documentation
sphinx>=4.1.2
sphinx-rtd-theme>=0.5.2
autodoc>=0.5.0

# Async support
asyncio>=3.4.3
aiofiles>=0.7.0

# System integration
psutil>=5.8.0          # For system metrics
pywin32>=301; sys_platform == 'win32'  # Windows system calls
wmi>=1.5.1; sys_platform == 'win32'    # Windows WMI interface

# Optional but recommended
rich>=10.6.0           # For beautiful terminal output
tqdm>=4.62.0           # For progress bars
python-dotenv>=0.19.0  # For environment variable management
